Adding more functionalities for clustering
now we provide an additional information in the table_all of the export_findmuon_data,
which is the clustering group of each site. in this we can understand better the clustering
that was done in the workflow.
the magnetic inequivalent should be then automatically appended as site with no cluster (or, the
cluster is just that site)